Working on Login page and Establishment configuration

+++ app/Http/Livewire/ActionsTable.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Livewire;
 
+use App\Models\Mail;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Laracasts\Flash\Flash;
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Views\Column;
@@ -9,12 +11,15 @@ use App\Models\Action;
 
 class ActionsTable extends DataTableComponent
 {
+    use AuthorizesRequests;
     protected $model = Action::class;
 
     protected $listeners = ['deleteRecord' => 'deleteRecord'];
 
     public function deleteRecord($id)
     {
+        $this->authorize('delete', Action::class);
+
         Action::find($id)->delete();
         Flash::success(__('messages.deleted', ['model' => __('models/actions.singular')]));
         $this->emit('refreshDatatable');
